+++ get_next_line/libft/includes/librb.h
@@ -33,12 +33,12 @@ typedef struct			s_node
 
 typedef struct  s_cont
 {
-    int     fd; 
-    char    *str;
-	char	*join;
-    int		end;
-	int		tmp;
-	int		len;
+    int			fd; 
+    char		*str;
+	char		*join;
+    size_t		end;
+	size_t		tmp;
+	size_t		len;
 }               t_cont;
 
 t_node					*ft_create_node(void *data);
